+++ src/css/deprecated-css-custom-property-syntax.ts
@@ -76,10 +76,10 @@ class DeprecatedCustomPropertySyntax extends CssRule {
   private addVarSyntaxWarnings(
       node: shady.Node, parsedDocument: ParsedCssDocument,
       warnings: Warning[]): any {
-    if (node.type === 'expression') {
+    if (node.type === shady.nodeType.expression) {
       const match = node.text.match(
           /var\s*\(\s*--[a-zA-Z0-9_-]+\s*,\s*(--[a-zA-Z0-9_-]+)\s*\)/);
-      if (match) {
+      if (match !== null) {
         const offsetOfVarInsideExpression = match.index!;
         const offsetOfSecondCustomPropWithinVar =
             match[0].match(/--[a-zA-Z0-9_-]+\s*\)$/)!.index!;